fixed core tests and started implementing update of observable classes.

+++ src/lib/core/profileManager.h
@@ -38,7 +38,7 @@
 
 const char DEFAULT_PROFILE_FILENAME[] = "presage.xml";
 
-const std::string  DEFAULT_LOGGER_LEVEL                 = "DEBUG";
+const std::string  DEFAULT_LOGGER_LEVEL                = "DEBUG";
 const int          DEFAULT_PREDICT_TIME                = 1000;
 const int          DEFAULT_MAX_PARTIAL_PREDICTION_SIZE = 100;
 const std::string  DEFAULT_COMBINATION_POLICY          = "Meritocracy";
@@ -46,7 +46,7 @@ const std::string  DEFAULT_PREDICTIVE_PLUGINS          = "";
 const int          DEFAULT_SLIDING_WINDOW_SIZE         = 80;
 const size_t       DEFAULT_SUGGESTIONS                 = 6;
 const bool         DEFAULT_REPEAT_SUGGESTION           = false;
-const unsigned int DEFAULT_GREEDY_SUGGESTION_THRESHOLD = 0;
+const size_t       DEFAULT_GREEDY_SUGGESTION_THRESHOLD = 0;
 const std::string  DEFAULT_PLUGINS                     = "";
 
 /** Juggles configuration files and presage system initialization.
